# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: ketuk-api
  version: 0.2.0
  description: |
    Backend Ketuk. Sukses JSON `{"data": ...}` kecuali auth (`data` + `tokens`) dan callback Duitku (teks `SUCCESS`).
    Schema katalog + undangan + planner sesuai handler hidup. CRUD lokasi/tamu/buku ucapan sudah dikode sesuai `docs/contracts/undangan-crud.md`.
    PATCH inviter dan caption galeri juga sudah dikode sesuai `docs/contracts/undangan-patch-gallery-inviter.md`.
    CRUD planner (patch/archive, hari, kategori, item budget, checklist) sudah dikode sesuai `docs/contracts/planner-crud.md` bagian 1; rundown sesi jam+PIC belum.
    Auth dan billing masih path-only. Vendor belum diimplementasi.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/auth/password:
    post:
      summary: Ganti password; semua sesi refresh dicabut
      security: [{bearerAuth: []}]
  /v1/me:
    get:
      summary: Profil, subscription, undangan, planner
      security: [{bearerAuth: []}]
    patch:
      summary: Ubah nama profil
      security: [{bearerAuth: []}]
  /v1/event-types:
    get:
      tags: [Catalog]
      operationId: listEventTypes
      summary: Katalog tipe acara
      description: "`invitation_schema` adalah petunjuk slot `content` untuk FE. Server tidak memvalidasi key di `content`."
      responses:
        "200":
          description: Daftar tipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventTypeListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/templates:
    get:
      tags: [Catalog]
      operationId: listTemplates
      summary: Template undangan
      description: "Publik. Bearer opsional — jika JWT valid dan paket punya `premium_templates`, template premium `locked=false`."
      parameters:
        - in: query
          name: type
          description: "Filter slug `event_types` (wedding, birthday, …)"
          schema: {type: string}
      responses:
        "200":
          description: Daftar template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/billing/subscription:
    get:
      summary: Paket aktif + entitlement
      security: [{bearerAuth: []}]
  /v1/billing/checkout:
    post:
      summary: Buat pesanan upgrade plan via Duitku
      security: [{bearerAuth: []}]
  /v1/orders:
    get:
      summary: Riwayat transaksi host (pesanan)
      security: [{bearerAuth: []}]
      parameters:
        - in: query
          name: kind
          description: subscription | gift | vendor; kosong = semua
          schema: {type: string}
  /v1/orders/{id}/sync:
    post:
      summary: Tanya status ke Duitku untuk order pending (rekonsiliasi)
      security: [{bearerAuth: []}]
  /v1/invitations:
    get:
      tags: [Invitation]
      operationId: listInvitations
      summary: Daftar undangan non-archived milik host
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: "List. Slice kosong di Go bisa `data: null`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: createInvitation
      summary: Buat undangan draft
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: "Draft baru. `type_id` diambil dari template. `content` kosong jadi `{}`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/EntitlementDenied"
        "404":
          description: template tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: getInvitation
      summary: Detail undangan
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Invitation]
      operationId: patchInvitation
      summary: Ubah judul, content, atau tanggal
      description: "Field yang tidak dikirim tidak diubah. `content` hanya diubah jika key JSON hadir (bukan null)."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchInvitationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: archiveInvitation
      summary: Archive undangan (bukan hard delete)
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Diarsipkan
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/publish:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    post:
      tags: [Invitation]
      operationId: publishInvitation
      summary: Publish dan isi slug
      description: "Slug bukan jalur RSVP. Idempoten jika sudah published. Tamu memakai `/v1/public/rsvp/{token}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/template:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    patch:
      tags: [Invitation]
      operationId: setInvitationTemplate
      summary: Ganti template (ditolak setelah publish)
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetTemplateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/EntitlementDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Undangan sudah published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: {code: immutable, message: cannot change template after publish}
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/locations:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listLocations
      summary: Daftar lokasi
      description: "Ubah / hapus satu lokasi: `/v1/invitations/{id}/locations/{locationId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addLocation
      summary: Tambah lokasi + maps_url
      description: "Tempel URL peta. Server tidak geocode. `id` di body diabaikan."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Location"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/locations/{locationId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/LocationId"
    patch:
      tags: [Invitation]
      operationId: patchLocation
      summary: Ubah lokasi
      description: "Body parsial: field yang tidak dikirim tidak diubah. `id` di body diabaikan. Server tidak geocode `maps_url`."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchLocationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteLocation
      summary: Hapus lokasi
      description: "Lokasi bukan milik undangan ini → 404 `location`."
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/guests:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listGuests
      summary: Daftar tamu + rsvp_token
      description: "Ubah / hapus satu tamu: `/v1/invitations/{id}/guests/{guestId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addGuest
      summary: Tambah tamu
      description: "Status awal `draft`. `rsvp_token` UUID baru untuk `/v1/public/rsvp/{token}`."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/EntitlementDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/guests/{guestId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/GuestId"
    patch:
      tags: [Invitation]
      operationId: patchGuest
      summary: Ubah tamu atau tandai invited
      description: "Body parsial. `status` host hanya `draft`|`invited` — jawaban RSVP tamu tidak lewat sini. `rsvp_token` tetap. Tidak mengirim notifikasi."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchGuestRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteGuest
      summary: Hapus tamu
      description: "Tautan RSVP tamu ikut mati. Kuota tamu menghitung baris, jadi slot kembali terpakai."
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/inviters:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listInviters
      summary: Turut mengundang (bukan tamu RSVP)
      description: "Ubah satu pengundang: `/v1/invitations/{id}/inviters/{inviterId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviterListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addInviter
      summary: Tambah yang turut mengundang
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/inviters/{inviterId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/InviterId"
    patch:
      tags: [Invitation]
      operationId: patchInviter
      summary: Ubah turut mengundang
      description: "Body parsial. `name` jika dikirim tidak boleh kosong setelah trim. `role` kosong boleh (hapus peran). Body `{}` = no-op."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchInviterRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteInviter
      summary: Hapus turut mengundang
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gallery:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listGallery
      summary: Daftar foto galeri
      description: "Ubah caption satu foto: `/v1/invitations/{id}/gallery/{mediaId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addGallery
      summary: Upload foto galeri
      description: "`multipart/form-data` field `file` (wajib), `caption` opsional. Content-Type harus `image/*`. Maks 16 MiB."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                caption:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gallery/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/MediaId"
    patch:
      tags: [Invitation]
      operationId: patchGallery
      summary: Ubah caption foto galeri
      description: "Hanya `caption`. File di `./storage` tidak disentuh. Id musik → 404 `media`. Body `{}` = no-op; `caption` kosong = hapus teks."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchGalleryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteGallery
      summary: Hapus foto galeri
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/music:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: getMusic
      summary: Musik undangan
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: "`data` null jika belum ada musik"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [Invitation]
      operationId: setMusic
      summary: Upload musik (mengganti yang lama)
      description: "Field `file`. Content-Type harus `audio/*`. Maks 16 MiB."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteMusic
      summary: Hapus musik
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus, atau tidak ada musik
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gift-methods:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listGiftMethods
      summary: Metode amplop (bank, ewallet, duitku)
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftMethodListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addGiftMethod
      summary: Tambah metode amplop
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGiftMethodRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftMethodResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gift-methods/{methodId}:
    delete:
      tags: [Invitation]
      operationId: deleteGiftMethod
      summary: Hapus metode amplop
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/InvitationId"
        - $ref: "#/components/parameters/GiftMethodId"
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/links:
    get:
      summary: List tautan; filter opsional invitation_id / planner_id
      security: [{bearerAuth: []}]
      parameters:
        - in: query
          name: invitation_id
          schema: {type: string, format: uuid}
        - in: query
          name: planner_id
          schema: {type: string, format: uuid}
    post:
      security: [{bearerAuth: []}]
  /v1/links/{id}:
    delete:
      summary: Hapus tautan (undangan dan planner tetap ada)
      security: [{bearerAuth: []}]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    InvitationId:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
...
    InternalError:
      description: Kesalahan server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: internal, message: internal server error}
  schemas:
...
    EventTypeListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/EventType"}
          nullable: true
    TemplateListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Template"}
          nullable: true
...
    InvitationListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Invitation"}
          nullable: true
...